Fix - handle duplication of Task name
If Task name is explicitly set in template, it duplicated
it here. Task name doesnt need to be in template, but by standard it should be in subset name.
This whole replace situation is because of avalon's Creator which modify subset name even if it shouldn't.
If Creator app is reworked (could have wide impact!), this should be cleaned up.