Keep `publisher.create_widget` variant when creating subsets
Whenever a person is creating a subset to publish, the "creator" widget
resets (where you choose the variant, product, etc.) so if the person
is publishing several images of the a variant which is not the default
one, they have to keep selecting the correct one after every "create".

This commit resets the original variant upon successful creation of a
subset for publishing.