Attempt to cleanup PreLoadJob situation
Most likely not viable, we want to support passing P4 credentials via environment variables, set either on machines or in Settings, but for that GlobalJobPreLoad would have to be called first, which it isn't.
Relative imports in DL are also major pita.
Kept here for posterity, but will be reverted immediately.